+++ Jarvis.Framework.Kernel/Support/GuaranteedDeliveryBroadcastBlock.cs
@@ -1,5 +1,6 @@
-﻿using Jarvis.Framework.Shared.Exceptions;
-using Metrics;
+﻿using App.Metrics;
+using Jarvis.Framework.Shared.Exceptions;
+using Jarvis.Framework.Shared.Support;
 using System;
 using System.Collections.Generic;
 using System.Threading;
@@ -19,7 +20,7 @@ namespace Jarvis.Framework.Kernel.Support
             {
                 options.BoundedCapacity = boundedCapacity;
             }
-            var meter = Metric.Meter($"GuaranteedDeliveryBroadcastBlock-{commitPollingClientId}", Unit.Items);
+            var meter = Metric.Meter($"GuaranteedDeliveryBroadcastBlock-{commitPollingClientId}", Unit.Items, TimeUnit.Milliseconds);
             var actionBlock = new ActionBlock<T>(
                 async item =>
                 {